:sparkles: Add basic routing
