+++ src/pages/DetailsVue.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>
+    <div>This is the details page</div>
+
+    <router-link to="/">Back</router-link>
+  </div>
+</template>
+
+<script>
+export default {}
+</script>
+
+<style></style>
